+++ decidim-core/app/controllers/concerns/decidim/action_authorization.rb
@@ -26,6 +26,8 @@ module Decidim
         "#{action}-#{resource.component.id}-#{resource.resource_manifest.name}-#{resource.id}"
       elsif resource && permissions_holder
         "#{action}-#{permissions_holder.class.name}-#{permissions_holder.id}-#{resource.resource_manifest.name}-#{resource.id}"
+      elsif permissions_holder
+        "#{action}-#{permissions_holder.class.name}-#{permissions_holder.id}"
       else
         "#{action}-#{current_component.id}"
       end